fix: bill edge banding by exact linear meters instead of rounding up
Edge-banding cost was rounded up to the whole meter (math.ceil(net + waste)),
which silently overcharged quotes and diverged from the raw totals shown
elsewhere (proforma workshop summary, public review link). It's now billed
exactly as net length + waste factor, with no rounding, so every view agrees
on the same figure.

quantity moves from int to float on the affected schemas/DB column
(order_lines.quantity) to hold the fractional linear meters; includes a
migration for the type change.

+++ src/modules/orders/model.py
@@ -314,13 +314,14 @@ class OrderLineModel(TimestampMixin, AuditMixin, Base):
     )
     product_code: Mapped[Optional[str]] = mapped_column(String(32), nullable=True)
     product_name: Mapped[Optional[str]] = mapped_column(String(128), nullable=True)
-    quantity: Mapped[int] = mapped_column(Integer)
+    # Boards: whole units. Edge banding: net linear meters + waste factor, billed
+    # exactly (no rounding up to a whole meter) — same value as ``linear_m``.
+    quantity: Mapped[float] = mapped_column(Float)
     unit_price_snapshot: Mapped[float] = mapped_column(Float)
     line_total: Mapped[float] = mapped_column(Float)
     avg_efficiency: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
     total_area_m2: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    # Edge banding: exact linear meters (incl. waste) for display; ``quantity``
-    # stores the whole meters billed. Null for boards.
+    # Edge banding: linear meters incl. waste (mirrors ``quantity``). Null for boards.
     linear_m: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
     # Half board: the line was charged at half (width/2, cost/2). False for
     # full boards and edge banding.
